NEWT: Add optional eager native initialization ; Proper exception handling
NEWT: Add optional eager native initialization
of Display and Screen to overcome a possible chicken/egg situation.
This is useful to be able to request the AbstractGraphicsDevice, via getGraphicsDevice().
Otherwise the abstract device won't be available before
the dependent components (Screen and Window) are realized.

Throw NativeWindowException in case native creation failed.