Adding NVIDIA 'Threaded optimization' workaround/fix at initialization on Windows for javaws/applets.
It has been observed that for some combinations, eg:
     - Windows 7 64bit (other variants may apply too)
     - NVIDIA 8600M GT
     - 260.99

the NVIDIA setting of 'Threaded optimization' := 'auto' (default) causes the JVM to simply crash
in case of javaws and [jnlp] applets.

'Threaded Optimization' := 'off' works reliable

'Threaded Optimization' := 'on'  never works with javaws and applets on the above configuration

A user could workaround this by setting 'Threaded Optimization' := 'off',
however, this would disable many users on the spot,
since you cannot ask the average user for such a task, if she only wants to see a web page.

This patch 'fixes' the 'auto' mode by running the eager GL profile initialization
within a block of single CPU affinity:

  SetProcessAffinityMask(pid, 1);
  try {
    initProfilesForDeviceImpl(device);
  } finally {
    SetProcessAffinityMask(pid, sysValue);
  }

Hopefully we can remove this hack with a driver fix.
However this workaround is as little invasive as possible.