Added two simple SWT unit tests.
Added a test that draws one triangle, using both the SWT
canvas and the AWT canvas with the SWT_AWT bridge. Also
added the SWT JARs for each platform to make/lib (since
that's where antlr.jar and junit.jar were stored). Modified
the make files to build and run the new tests.