Fix bug 472: Eclipse classpath not updated ...
An excluded package was recently moved.  The Eclipse ".classpath"
file should be updated to reflect the change.  Currently Eclipse
will not compile the entire project because the package has
errors.  Building on the command line using the Ant build file
seems to be fine however.

The package was

    com/jogamp/opengl/impl/gl2/fixme

but is now

    jogamp/opengl/gl2/fixme