Prevent native libraries from always rebuilding and stripping
The cc task will always re-link, even if no source files are built.
This commit wraps an uptodate task around cc to prevent this, so
the libraries will only be rebuilt now if a source file changes.

This commit also moves library symbol stripping into the same
uptodate task, so it only happens if a library is built (previously
the libraries were always stripped).