UnitTests AWT: Remove GLProfile.initSingleton(true) and ensure AWT-dispose happens on AWT-EDT
We will use the default implicit call of GLProfile.initSingleton(false).
(This is the same case as for AWT applets as.)

We shall create extra test cases for AWT + GLProfile.initSingleton(true)
to test it's stability. However .. to nail down our test instability w/ AMD's fglrx driver
we move to the default behavior.