NEWTWindow Focus: Skip requestFocus() if already owning focus; setFullscreen() requests focus 'later'.
We shall rely on the focus state, hence we can skip 'requestFocus()'
if we already own the focus. This allows a fast-path especially when called
from native code (mouse click).

Request focus 'later' on setFullscreen() allowing native WM to complete event handling,
this is required especially on X11 to guarantee a focused fullscreen window.