NEWT/X11: Fix Insets determination for undecorated / child window.
Due to latest changes, a bug is disposed where the native 'updateInsets'
attempts to determine the window insets by it's parent window (fall-back).
The latter works only in case of a top-level window.

Adding query to the WM whether the window is decorated (top-level) or not.
Attempt to use the fall-back parent window method in case of a decorated window.

This whole inset code based on the parent window is probably completly redundant,
nevertheless we keep it alive until further notice.