Adding initial Libav/FFMpeg GLMediaPlayer implementation
The Java classes already slipped through in
commit 10935e1ec0d8ed677bc3fddfaa8cd73898a3bcbf - oops.

Since we cannot provide a Libav binary (even though Google does in Android and Chrome)
due to legal uncertainities .. we dynamically link to an existing Libav / FFmpeg library
in a 'relaxed' manner. Ie. we allow certain recent functions to be absent to be able
to run on a wider range of Libav versions.

Currently tested on Debian Linux and Windows7 64bit/32bit
Binaries for Win/OSX:
  - Windows http://ffmpeg.zeranoe.com/builds/
  - OSX http://www.ffmpegx.com/

Features:
  - Dynamic relaxed linking to Libav (see above)
  - YUV420P texture lookup function shader stub (conversion to RGB)
  - 1-copy only (decoder buffer to texture)
  - simple
  - uses libavformat's network streaming
  - fixes some odd PTS values

TODO:
  - Audio output (Should use OpenAL, duh)
  - Seek works poorly
  - Offthread multi-texture fetching for smoother animation
  - Maybe more pixelformat conversions