Bug 556: Newt Mouse Synthetic Drag Event: Clear state if mouse enters/leaves window.
Since we cannot guarantee to have the pressed button information when receiving
the mouse move event, we synthesize the dragged event (move while mouse button pressed).

To simplify the situation and have a compromise,
we clear the mouse pressed states when mouse
enters or leaves the window to remove the dragged events at re-entering.

This seems more sensible, since dragging after re-entering the mouse shall not be expected.