Fix Bug 617 (part2): Workaround buggy GPU driver where FBO reattachment of Colorbuffer w/ diff size leads to GL_FRAMEBUFFER_UNSUPPORTED.
This occured at least on:
  - OS X 10.6.8
  - GL_RENDERER    NVIDIA GeForce 7300 GT OpenGL Engine
  - GL_VERSION     2.1 NVIDIA-1.6.36

Remedy is to catch the exception @ GLFBODrawableImpl.reset(..) and switch over to
fallback 'reset' method:
  FBO reattachment -> FBO complete recreation

Of course, the FBO recreation is noticable slower,
but at least it seems to work on the offending system.

Not tested on the offending system, but manually provoked GLException on FBOObject
to trigger fallback, which is working here.