Fix for JOGL bug 629, and added new unit tests.
The change to AWTNewtEventFactory appears to fix the original issue
for AWTCanvas instances, and the TestNewtEventModifiersAWTCanvas
appears to work ok too.  However, there are still issues with
NewtCanvasAWT and NewtCanvasSWT instances.  These might be problems
in the test code, but there's also a good chance there are still
issues in the NEWT event delivery infrastructure.  For the time being
I recommend that only TestNewtEventModifiersAWTCanvas be included
in routine unit tests.

The tests are defined in TestNewtEventModifiers, and the remaining
test classes extend it to define how the window and associated
GL drawing surface are created.

File ModifierMappings.txt is simply informational, and shows how
the modifier bits are laid out between AWT and NEWT.  This possibly
should have been a spreadsheet.