Enhance PMVMatrix: Use FloatStack to implement PushMatrix and PopMatrix (Bug 657)
One FloatStack for each, MV, P and T, with initialialSize zero to save memore
and growSizes: MV = 16 arrays, P = 2 arrays, T = 2 arrays,
with array = 16 floats.

This shall save performance due to the preallocated stack when used
and growing only in the above mentioned grow intervals.

We may finetune the growSize, if required.