Bug #677 : fix
There is no way to know the visibility of
the softKeyboard on Android. So, we expect that
the developer uses GLWindow::setKeyboardVisible()
and store the visibility.

About comment #3 : we want the regular Back button
behavior when the softKeyboard is visible,
so dispatchKeyEventPreIme() must return false
and let the IME handle the event.