Support specifying the X11 default display with a system property
Usually, the environment variable DISPLAY is used to define the default
display under X11. There are cases where the default display has to be
defined or overridden programmatically. For those cases, the new system
property nativewindow.x11.display.default can be used (as environment variables can't
be changed in Java). Its semantics are identical to the DISPLAY
environment variable.