FFMPEGMediaPlayer: Handle v-flipped 'bottom-up' pictures ; Refine API doc 'camera ID'
If linesize is < 0, it is not invalid as assumed in commit eca6a5cb1e2beda84dfbafc31ed225e272f4f3fb,
but vertically flipped (bottom-up).

We have to adjust the data pointers, which are moved to the upper end of memory as well
and can proceed as usual.

TODO:
  - Update texture 'mustFlipVertically' to 'false' in this case.
  - Later:
    - Allow updating texture size ..
    - Whole pixel-fmt/texture-lookup-shader association must scale better,
      i.e. extract the 'knowledge' into one class, use a static shader code
      using uniforms instead of hard-coded values .. etc.