Fix Bug 816: JAWTWindow.getLocationOnScreenNonBlocking(..) shall use JRootPane for last position offset; AWTMisc.getInsets(..) add special JRootPane case.
Please note that we use JAWTWindow.getLocationOnScreenNonBlocking(..) to determine the
location on the screen 'only' b/c we cannot allow AWT to aquire the tree-lock!
The latter would be the case if using AWT's 'getLocationOnScreen()'.

If anybody has a more reliable implementation to achieve the same .. please provide your patch!

The following fix has been performed to fix the last issue w/ vZome.

- JAWTWindow.getLocationOnScreenNonBlocking(..) shall use JRootPane for last position offset
  - w/ vZome, the frame's position was 0/0 (invalid), instead JRootPane's position is good.
    Use JRootPane's position and stop traversing here (LW top-level).

- AWTMisc.getInsets(..) add special JRootPane case.
+     * Exception is JRootPane.
+     * Return it's parent's Window component's insets if available,
+     * otherwise return JRootPane's insets.<br>
+     * This is due to <i>experience</i> that <i>some</i> JRootPane's
+     * do not expose valid insets value.
  -