Bug 860 - AWT Printing (AWTTilePainter): Shall use the enclosing integer rectangle of the scaled double precision clipping rect
AWTTilePainter simply rounds the scaled double precision clipping rectangle
to receive the integer rectangle.

This leads to uncovered drawing areas, since the integer rectangle position
could be greater - and the size could be smaller than the double precision source.

To get the enclosing rectangle, we need to use
  iPos = floor(position)
  iSize = ceil(position+size) - floor(position)

.. turns our that the double precision 'Rectangle Rectangle2D.getBounds()'
already performs this math.