Bug 975 - GLJPanel's OffscreenDrawable shall not double swap - Refine unit test for visual validation of 'no frame lag'
To validate whether a 'display' command w/o animator results to the desired frame
we introduce a 'userCounter' in TextRendererGLEL.

The latter gets increased and maybe visually validated by a key-press -> display.

Results: In all modes, MSAA or !MSAA, or flip - the result is valid.

Tested on windows and linux.