Copyright: Add missing Copyright tag
Adding the missing JogAmp copyright tag on certain files.

Currently in debate whether 'JogAmp Community' is a legal Copyright tag in the first place,
we might need to add (or replace it with) my authorship.
However, as authorship is well documented via the git repository,
this should be no real world issue.